+++ InvestableIndexCore/IndexData.h
@@ -17,7 +17,7 @@ namespace InvestableIndex {
 		std::vector<double> m_CumReturn;
 
 	private:
-		int getValues(long long date, std::vector<long long>* stk, std::vector<double>* openvalue, std::vector<double>* closevalue, StkConContext& ctx) const;
+		long long getValues(long long date, std::vector<long long>* stk, std::vector<double>* openvalue, std::vector<double>* closevalue, StkConContext& ctx) const;
 
 	public:
 		IndexData() : m_dataset(nullptr), m_params(nullptr) {}
@@ -31,9 +31,9 @@ namespace InvestableIndex {
 		IndexData& setParams(IndexParameters& params) { m_params = &params; return *this; }
 		IndexParameters& getParams() { return *m_params; }
 
-		int getWeight(long long date, std::vector<long long>* stk, std::vector<double>* weight, int type) const;
-		int getOpenWeight(long long date, std::vector<long long>* stk, std::vector<double>* weight) const { return getWeight(date, stk, weight, OPENPRICECOL); }
-		int getCloseWeight(long long date, std::vector<long long>* stk, std::vector<double>* weight) const { return getWeight(date, stk, weight, CLOSEPRICECOL); }
+		long long getWeight(long long date, std::vector<long long>* stk, std::vector<double>* weight, int type) const;
+		long long getOpenWeight(long long date, std::vector<long long>* stk, std::vector<double>* weight) const { return getWeight(date, stk, weight, OPENPRICECOL); }
+		long long getCloseWeight(long long date, std::vector<long long>* stk, std::vector<double>* weight) const { return getWeight(date, stk, weight, CLOSEPRICECOL); }
 
 		const std::vector<double>& getDailyReturn() const { return m_DailyReturn; }
 		const std::vector<double>& getCumReturn() const { return m_CumReturn; }
