+++ InvestableIndexCore/SizeTopPool.h
@@ -15,13 +15,13 @@ namespace InvestableIndex {
 		const DataSet* m_DataSet = nullptr;
 		const IndexParameters* m_Params = nullptr;
 		WeightType m_WeightType = WeightType::FREESHARE;
-		int m_TopCount = 50;
+		long long m_TopCount = 50;
 	public:
 		SizeTopPool() {}
 		~SizeTopPool() {}
 
 		void init(const StkPool& base, const DataSet& data, const IndexParameters& params) { m_Pool = &base; m_DataSet = &data; m_Params = &params; }
-		StkPool& setTopCount(int topcnt) { m_TopCount = topcnt; return *this; }
+		StkPool& setTopCount(long long topcnt) { m_TopCount = topcnt; return *this; }
 		StkPool& setPool(StkPool& pool, WeightType wt) { m_Pool = &pool; m_WeightType = wt; return *this; }
 
 		virtual long long getSnapshot(long long date, std::vector<long long>* list) const;
